Adding loader to gltf
