+++ app/components/canvas/Scene.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import { Canvas } from '@react-three/fiber'
+import { Preload } from '@react-three/drei'
+import { r3f } from '@/app/helpers/global'
+
+export default function Scene({ ...props }) {
+  return (
+    <Canvas {...props}>
+      <r3f.Out />
+      <Preload all />
+    </Canvas>
+  )
+}
